+++ packages/expo-updates/ios/EXUpdates/UpdatesStateMachine.swift
@@ -147,6 +147,11 @@ internal struct UpdatesStateEventDownloadComplete: UpdatesStateEvent {
 }
 
 /**
+ Date formatter for the last check times sent in JS events
+ */
+let iso8601DateFormatter = ISO8601DateFormatter()
+
+/**
  The state machine context, with information that will be readable from JS.
  */
 internal struct UpdatesStateContext {
@@ -160,6 +165,14 @@ internal struct UpdatesStateContext {
   let downloadedManifest: [String: Any]?
   let checkError: [String: String]?
   let downloadError: [String: String]?
+  let lastCheckForUpdateTime: Date?
+
+  private var lastCheckForUpdateTimeDateString: String? {
+    guard let lastCheckForUpdateTime = lastCheckForUpdateTime else {
+      return nil
+    }
+    return iso8601DateFormatter.string(from: lastCheckForUpdateTime)
+  }
 
   var json: [String: Any?] {
     return [
@@ -172,7 +185,8 @@ internal struct UpdatesStateContext {
       "latestManifest": self.latestManifest,
       "downloadedManifest": self.downloadedManifest,
       "checkError": self.checkError,
-      "downloadError": self.downloadError
+      "downloadError": self.downloadError,
+      "lastCheckForUpdateTimeString": lastCheckForUpdateTimeDateString
     ] as [String: Any?]
   }
 }
@@ -189,6 +203,7 @@ extension UpdatesStateContext {
     self.downloadedManifest = nil
     self.checkError = nil
     self.downloadError = nil
+    self.lastCheckForUpdateTime = nil
   }
 
   // struct copy, lets you overwrite specific variables retaining the value of the rest
@@ -210,6 +225,7 @@ extension UpdatesStateContext {
     var downloadedManifest: [String: Any]?
     var checkError: [String: String]?
     var downloadError: [String: String]?
+    var lastCheckForUpdateTime: Date?
 
     fileprivate init(original: UpdatesStateContext) {
       self.isUpdateAvailable = original.isUpdateAvailable
@@ -222,6 +238,7 @@ extension UpdatesStateContext {
       self.downloadedManifest = original.downloadedManifest
       self.checkError = original.checkError
       self.downloadError = original.downloadError
+      self.lastCheckForUpdateTime = original.lastCheckForUpdateTime
     }
 
     fileprivate func toContext() -> UpdatesStateContext {
@@ -235,7 +252,8 @@ extension UpdatesStateContext {
         latestManifest: latestManifest,
         downloadedManifest: downloadedManifest,
         checkError: checkError,
-        downloadError: downloadError
+        downloadError: downloadError,
+        lastCheckForUpdateTime: lastCheckForUpdateTime
       )
     }
   }
@@ -334,6 +352,7 @@ internal class UpdatesStateMachine {
         $0.latestManifest = nil
         $0.isUpdateAvailable = false
         $0.isRollback = false
+        $0.lastCheckForUpdateTime = Date()
       }
     case .checkCompleteAvailable:
       return context.copy {
@@ -342,11 +361,13 @@ internal class UpdatesStateMachine {
         $0.latestManifest = event.manifest
         $0.isRollback = event.isRollback
         $0.isUpdateAvailable = true
+        $0.lastCheckForUpdateTime = Date()
       }
     case .checkError:
       return context.copy {
         $0.isChecking = false
         $0.checkError = event.error
+        $0.lastCheckForUpdateTime = Date()
       }
     case .download:
       return context.copy {